Add cycle counting
The CPU may spend an extra cycle fixing the upper address when adding an
offset, depending on the type of instruction (as indicated by
InstructionType) and whether a page boundary was crossed. Read
instructions tend to only require that extra cycle if a page was
crossed, while all other instructions will always use that extra cycle.
This page cross detection and cycle increments are now performed by
relativeAddress().
As address modes are processed in getTarget(), this function is now
provided with the type of instruction it is processing, so cycle counts
can be calculated correctly.

Popping from the stack requires spending one cycle beforehand, to
pre-increment the stack pointer. Subsequent pops within an instruction
only require a single cycle, as the stack pointer can be incremented in
the same cycle as popping from the stack.
The preIncrement argument to pop() and popFlags() must be set to true
for the first pop operation of the instruction to include this initial
cycle.

read() is no longer [[nodiscard]] as some instructions will implicitly read
and discard the result. Such reads have been annotated with a comment.

Some instructions seem to always use an extra cycle during execution, so
in those cases the cycle is manually incremented within the instruction.